[dv,ralgen] update `ralgen.py` to use git paths over relative
The `ralgen.py` script was using relative paths to find the `topgen.py`
script. This caused issues when migrating `dvsim.py` to use bazel over
meson, specifically two consecutive invocations of `dvsim.py` without a
`bazel clean` in between seemed to cause the `topgen.py` not to be
found.

This updates the `ralgen.py` script to use `git` to resolve absolute
paths to tools in the repo, which fixes the above issue.

Signed-off-by: Timothy Trippel <ttrippel@google.com>