[otbn,dv] Improve FIPS connection for RND in OTBN
Implements a change in how we use EDN agent. Now we are polling
for a request from OTBN to generate a random data with controlled
randomness for its FIPS bit. By default we will always randomize
FIPS bit as 1.

By overwriting the `rnd_fips_pct` variable from
the config, we can allow agent to generate FIPS bits with more
flexibility as well.

Signed-off-by: Canberk Topal <ctopal@lowrisc.org>