[usbdev] Remove internal CDC logic
With the CSR clock now required to be the 48 MHz USB clock, there is no
domain crossing between the TL-UL side and the internals (within the
module). Instead, the clock domain crossing is outside the module (in
xbar_main, in this case).

Also change AV FIFO empty / RX FIFO full interrupts to reflect the
status, instead of waiting for an underflow / overflow event.

Signed-off-by: Alexander Williams <awill@google.com>