[dv/otp_ctrl] Fix lc_dft_en failures
This PR fixes the regression failures when lc_dft_en is set to On and DV
did not expect it to return d_error.
This case is hard to align on the scoreboard side because it is timing
sensetive regarding the d_channel and a_channel.
Because this case is not likely to happen in chip level (lc_dft_en
cannot be switched on and off on the fly), I added a TODO to support
checking that in otp_ctrl_test_access sequence.

Signed-off-by: Cindy Chen <chencindy@opentitan.org>