[dv/chip] Fix regression error on lc_ctrl_transition test
This PR fixes the regression failure on lc_ctrl_transition_test where
the testbench expects an external clock switch (PR #13141). However, the
default setting for lc_ctrl_transition test is to use internal clock. So
I fix it by adding a CSR read to check if the ctrl register is
configured to use internal or external clock.

Signed-off-by: Cindy Chen <chencindy@google.com>