[otbn,rtl,dv] Add status for internal secure wipe
After execution OTBN securely wipes its internal state.  Prior to this
commit, OTBN remained in the `BUSY_EXECUTE` status during this secure
wipe.  The `BUSY_EXECUTE` status, however, is defined as "OTBN is busy
executing software", which is not accurate: when performing a secure
wipe, OTBN no longer executes software.  Additionally, there is work in
progress to perform an internal secure wipe after reset as well as on a
software command (the latter is already possible for the instruction and
the data memories).  The newly added status will be used to represent
those operations as well.

Signed-off-by: Andreas Kurth <adk@lowrisc.org>