fix(kmac): Maintain Sideload Key
Issue #13655

Problem
-------

CSR Key is used incorrectly if:

1. App Intf is configured to KMAC mode
2. App Intf sends short message length.

Analysis
--------

`kmac_app` sends sideloaded key to `kmac_core` if:

1. !!CFG.sideload is set
2. Or, the mux selection is configured to `SelApp`

The issue here is, the logic does not maintain `mux_sel` value to
`SelApp` until the hashing operation is completed. Right after the FSM
moves to `StAppOutLen`, the `mux_sel` is changed to `SelOutLen` to send
the key length to `kmac_core`.

The FSM moves to `StAppOutLen` when the App Interface logic receives the
last message byte. If the message size is small, FSM may move to the
state before the `kmac_core` processes the sideloaded key in the SHA3
engine.

The key shall be maintained, at least, until the SHA3 processes the key
block. As the key may changed to the CSR key, incorrect key will be used
in this case.

Resolution
----------

`mux_sel*` signals are for data mux. For sideload keys, it is better to
follow `sha3_prefix_o` logic. The logic uses the FSM value to maintain
the PREFIX until the hashing op is completed.

Signed-off-by: Eunchan Kim <eunchan@opentitan.org>