[otbn,rtl] Fix `adder_y_res_used`
`adder_y_res_used` indicates whether `adder_y_res` is used in any way.
This is relevant for determining if upstream integrity errors could
possibly have any effect.  Prior to this commit, `adder_y_res_used` was
`0` in one case where `adder_y_res` is used in an `if` condition.  This
is wrong, and this commit fixes this flaw.

Signed-off-by: Andreas Kurth <adk@lowrisc.org>