[fpv/kmac] Fix shadow reg assertion error
When shadow reg assertion error happens, kmac entropy will go to FSM
error state and drop the entropy request without waiting for an entropy
response.
This will cause an assertion error. To avoid this error, I modify the
RTL to disable the assertion check when FSM error state is high.

Signed-off-by: Cindy Chen <chencindy@opentitan.org>