[entropy_src/dv] Account for SHA3 non-reset on disable
Unlike most of the other blocks within the entropy_src, the SHA3 block
is not cleared on a disable event. (The SHA3 requires particular
sequencing to be shut down gracefully, and the disable operation
is otherwise atomic).

This is not a security problem as there is no harm if some excess
entropy remains in the SHA3 state before starting the next seed.
However this does need to be modelled properly in the scoreboard.

This commit changes the process_fifo_q into a 64-bit fifo to
reflect the fact that the SHA block receives data in 64 bit
chunks.  This scoreboard FIFO is then not cleared during
continuous-mode disable events.

This SHA behavior is now properly modelled both in the default mode
as well as in the FW_OV_INSERT mode.

Signed-off-by: Martin Lueker-Boden <martin.lueker-boden@wdc.com>