[dv, xcelium] Indicate SVA-disabled hierarchies
Xcelium sim throws a runtime warning indicating that there are
several hierarchies in the design where asssertions are permanently
disabled. This change in the tcl file enables the dumping of those
list of hierarchies as an informational message (which gets rid of
the warning message). The list is benign - it only contains a bunch
of assertion cover statements.

Signed-off-by: Srikrishna Iyer <sriyer@google.com>