fix(hmac): Remove Unreachable Error
Issue https://github.com/lowRISC/opentitan/issues/3022

`SwPushMsgWhenDisable` is unreachable as `msg_allowed` signal already
protects the error scenario `msg_valid && !sha_en`.

This commit removes the error signal but leaves the Error code in
`hmac_pkg` to be compatible to the previous code. Soon, in the next
release, the code may be renamed to `DeprecatedSwPushMsgWhenDisable`.

Signed-off-by: Eunchan Kim <eunchan@opentitan.org>