[sigverify] Add an automatic check for OTBN instruction count ranges.
This adds a custom Bazel test that can run in CI to check the min/max
instruction count range in mod_exp_otbn.h against the values calculated
by the new static analysis script. Adding a test in this way seems
preferable to automatically generating the numbers, because we should in
general be double-checking the script's answers with manual analysis to
be extra certain.

Signed-off-by: Jade Philipoom <jadep@google.com>