[entropy_src/dv] Track FW_OV FIFO exceptions
This commit aims to support verification of the the new
recoverable ES_FW_OV_WR_ALERT by tracking FIFO overflow conditions
inside the DUT

To a achieve this a new entropy_subsys_fifo_exception_if class is
added.  This interface can be bound to either a packer FIFO or
a synchronous FIFO.  (Though these two primitives have some port
differences, the interface is parametrizeable to attach to either,
leaving unbound/non-existent ports unused).

This new FIFO exception interface is identified as specific to the
entropy subsystem, as only the IPs in this subsystem have exceptions
for these FIFO events.

This new interface is then bound to the Pre-conditioning packer
FIFO, and the scoreboard is updated to properly predict future
DUT outputs should one of these events occur.

Further more the FW_OV test, now runs on average at a higher rate,
in order to both test the new testbench functionality, but also
to generate more seeds for better coverage in each test.

Signed-off-by: Martin Lueker-Boden <martin.lueker-boden@wdc.com>